OAK-9783 - Excerpt Test ported to Elastic
